Lambda map.
The square operation is to be performed on an input array of arrays.
It is to be applied only to the integers which are greater than 0,
ignoring the rest, i.e., the rest are not to be included in the output.
The input is not exactly a 2D array as all the rows can have
a different number of elements.